+++ eclipse/plugins/org.switchyard.tools.models.switchyard1_0/src/org/switchyard/tools/models/switchyard1_0/soap/ProxyType.java
@@ -2,6 +2,7 @@
  */
 package org.switchyard.tools.models.switchyard1_0.soap;
 
+import java.math.BigInteger;
 import org.eclipse.soa.sca.sca1_1.model.sca.CommonExtensionBase;
 
 /**
@@ -21,7 +22,7 @@ import org.eclipse.soa.sca.sca1_1.model.sca.CommonExtensionBase;
  * </p>
  *
  * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType()
- * @model extendedMetaData="name='ProxyType' kind='empty' namespace='##targetNamespace'"
+ * @model extendedMetaData="name='ProxyType' kind='elementOnly'"
  * @generated
  */
 public interface ProxyType extends CommonExtensionBase {
@@ -33,6 +34,9 @@ public interface ProxyType extends CommonExtensionBase {
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
+     * <!-- begin-model-doc -->
+     * Proxy Type. Default is HTTP. Allowed values are HTTP and SOCKS
+     * <!-- end-model-doc -->
      * @return the value of the '<em>Type</em>' attribute.
      * @see #setType(String)
      * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType_Type()
@@ -60,6 +64,9 @@ public interface ProxyType extends CommonExtensionBase {
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
+     * <!-- begin-model-doc -->
+     * Proxy Host.
+     * <!-- end-model-doc -->
      * @return the value of the '<em>Host</em>' attribute.
      * @see #setHost(String)
      * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType_Host()
@@ -87,14 +94,17 @@ public interface ProxyType extends CommonExtensionBase {
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
+     * <!-- begin-model-doc -->
+     * Proxy Port.
+     * <!-- end-model-doc -->
      * @return the value of the '<em>Port</em>' attribute.
-     * @see #setPort(Integer)
+     * @see #setPort(BigInteger)
      * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType_Port()
-     * @model dataType="org.eclipse.emf.ecore.xml.type.IntObject"
+     * @model dataType="org.eclipse.emf.ecore.xml.type.Integer"
      *        extendedMetaData="kind='element' name='port' namespace='##targetNamespace'"
      * @generated
      */
-    Integer getPort();
+    BigInteger getPort();
 
     /**
      * Sets the value of the '{@link org.switchyard.tools.models.switchyard1_0.soap.ProxyType#getPort <em>Port</em>}' attribute.
@@ -104,7 +114,7 @@ public interface ProxyType extends CommonExtensionBase {
      * @see #getPort()
      * @generated
      */
-    void setPort(Integer value);
+    void setPort(BigInteger value);
 
     /**
      * Returns the value of the '<em><b>User</b></em>' attribute.
@@ -114,10 +124,13 @@ public interface ProxyType extends CommonExtensionBase {
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
+     * <!-- begin-model-doc -->
+     * Proxy user name.
+     * <!-- end-model-doc -->
      * @return the value of the '<em>User</em>' attribute.
      * @see #setUser(String)
      * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType_User()
-     * @model dataType="org.eclipse.emf.ecore.xml.type.String"
+     * @model dataType="org.eclipse.emf.ecore.xml.type.String" required="true"
      *        extendedMetaData="kind='element' name='user' namespace='##targetNamespace'"
      * @generated
      */
@@ -141,10 +154,13 @@ public interface ProxyType extends CommonExtensionBase {
      * there really should be more of a description here...
      * </p>
      * <!-- end-user-doc -->
+     * <!-- begin-model-doc -->
+     * Proxy password.
+     * <!-- end-model-doc -->
      * @return the value of the '<em>Password</em>' attribute.
      * @see #setPassword(String)
      * @see org.switchyard.tools.models.switchyard1_0.soap.SOAPPackage#getProxyType_Password()
-     * @model dataType="org.eclipse.emf.ecore.xml.type.String"
+     * @model dataType="org.eclipse.emf.ecore.xml.type.String" required="true"
      *        extendedMetaData="kind='element' name='password' namespace='##targetNamespace'"
      * @generated
      */
