+++ eclipse/plugins/org.switchyard.tools.models.switchyard1_0/src/org/switchyard/tools/models/switchyard1_0/soap/impl/ProxyTypeImpl.java
@@ -2,6 +2,7 @@
  */
 package org.switchyard.tools.models.switchyard1_0.soap.impl;
 
+import java.math.BigInteger;
 import org.eclipse.emf.common.notify.Notification;
 
 import org.eclipse.emf.ecore.EClass;
@@ -79,7 +80,7 @@ public class ProxyTypeImpl extends CommonExtensionBaseImpl implements ProxyType 
      * @generated
      * @ordered
      */
-    protected static final Integer PORT_EDEFAULT = null;
+    protected static final BigInteger PORT_EDEFAULT = null;
 
     /**
      * The cached value of the '{@link #getPort() <em>Port</em>}' attribute.
@@ -89,7 +90,7 @@ public class ProxyTypeImpl extends CommonExtensionBaseImpl implements ProxyType 
      * @generated
      * @ordered
      */
-    protected Integer port = PORT_EDEFAULT;
+    protected BigInteger port = PORT_EDEFAULT;
 
     /**
      * The default value of the '{@link #getUser() <em>User</em>}' attribute.
@@ -197,7 +198,7 @@ public class ProxyTypeImpl extends CommonExtensionBaseImpl implements ProxyType 
      * <!-- end-user-doc -->
      * @generated
      */
-    public Integer getPort() {
+    public BigInteger getPort() {
         return port;
     }
 
@@ -206,8 +207,8 @@ public class ProxyTypeImpl extends CommonExtensionBaseImpl implements ProxyType 
      * <!-- end-user-doc -->
      * @generated
      */
-    public void setPort(Integer newPort) {
-        Integer oldPort = port;
+    public void setPort(BigInteger newPort) {
+        BigInteger oldPort = port;
         port = newPort;
         if (eNotificationRequired())
             eNotify(new ENotificationImpl(this, Notification.SET, SOAPPackage.PROXY_TYPE__PORT, oldPort, port));
@@ -292,7 +293,7 @@ public class ProxyTypeImpl extends CommonExtensionBaseImpl implements ProxyType 
                 setHost((String)newValue);
                 return;
             case SOAPPackage.PROXY_TYPE__PORT:
-                setPort((Integer)newValue);
+                setPort((BigInteger)newValue);
                 return;
             case SOAPPackage.PROXY_TYPE__USER:
                 setUser((String)newValue);
